LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/aspect-ratio/replaced-element-042-expected.txt


FAIL canvas 1 assert_equals:
<canvas width="100" height="50" style="max-width: 50px; max-height: 70px; padding-top: 50px" data-expected-width="50" data-expected-height="50"></canvas>
width expected 50 but got 20
FAIL canvas 2 assert_equals:
<canvas width="50" height="100" style="max-width: 70px; max-height: 50px; padding-left: 50px" data-expected-width="50" data-expected-height="50"></canvas>
width expected 50 but got 70
FAIL canvas 3 assert_equals:
<canvas width="50" height="100" style="min-width: 150px; min-height: 175px; padding-left: 50px" data-expected-width="175" data-expected-height="175"></canvas>
width expected 175 but got 225
FAIL canvas 4 assert_equals:
<canvas width="100" height="50" style="min-width: 175px; min-height: 150px; padding-top: 50px" data-expected-width="175" data-expected-height="175"></canvas>
height expected 175 but got 225
